Fix heirachical dependency bug in depcheck
For some reason, some packages in centos5-rutgers are not in
centos5-rutgers-testing, but are dependencies for packages that are.
If centos5-rutgers-testing is enabled on a system, centos5-rutgers will
be as well. Thus, the stable repository must be enabled when checking
the testing, and both testing and stable must be enabled for unstable.